requirements.txt
lxml==4.2.5
requests==2.19.1
cfscrape==2.1.1
PyExecJS==1.5.1
retrying==1.3.3
PyQt5==5.15.4
